Add new window to display detailed internship view
MainWindow is the main UI of the application. It holds references to
the helpWindow, resultDisplay, InternshipListPanel and more.
Currently, it does not support a separate window for a more
detailed view of an internship.

To support this new feature, we need to show a new window that we
will populate with our selected internship for the detailed view.

Let's
* Add mouse click handlers to update the selected internship in Logic
* Add a new DetailedInternshipWindow that we can show/hide