Fix bug where deadline added does not reflect on UI
The current addDeadline command does not call
model.setInternship(). However, this is required to
basically trigger a UI refresh (via ObservableList
listener).

Let's add a line to set the internship that we added
the deadline to, to replace itself (no change to
internal internship list). This triggers the UI to
update which displays the added deadline.