Rewrite add and delete commands for User Guide
The biggest change here is the the removal of the screenshots for
successful output and error as this screenshots do not provide
much value at the moment without further context.

Instead, the users are now directed to the Tutorial section which
has a more contextualised use case for both Add and Delete.