Fix bug with detailed view window and edit
Currently, when an internship is being edited, if it is being
displayed on the detailed view window, it would dissapear on occasion.
This is discovered to be because the identity fields of Internship
are being used to check for equality (aka isSameInternship()).

It would be useful for users if the internship that is being viewed
does not dissapear upon any field update.

This is made possible with a simple change to edit command. By
checking if the internship to be edited is the model's current
selected internship (which is the one displayed in the GUI), the
post-edit internship can be set to the new selected internship
in model. (A non selected internship post-edit will NOT appear
in the detailed view window).

This is corroborated by 2 additional test cases in
InternshipEditCommandTest which checks that a selected internship
post edit will still be the selected internship, and a non-selected
internship will not become the selected internship.

Let's implement these changes within edit command.