Add tip describing filter reset upon execution of commands
Our current implement of model that sets the predicate to
show all internships upon execution of most commands causes
the find filter to be reset every time a command is executed.

There are a few exceptions like delete, sort etc. that are also
highlighted in this change.

Let's
* Make explicit this behaviour in the UG.

For future reference
* This feature will be under our planned enhancements.