Increase visibility of caution message for copy pasting commands
Currently, there is a warning to users about copy pasting commands
directly from the PDF version of the UG under the 'Notes about
command format' section.

This is not very visible to users as it is the 4th bullet point
and is likely to be skipped through.

Let's
* Improve it's visibility by wrapping in a 'Caution' formatting
* Place it at the beginning in 2 places, under `Tutorial` and
`Commands` sections since these are the places where the user
is most likely to copy paste commands.