Fix null invocation bug
Explanation of this bug:
This was created during the pull #197. This added a check for
the current user selected internship, which ASSUMES that there
would be an internship selected (defaulted to first internship
in list).

This is true when the app is just launched with sample data. However
when the data is cleared, there would be 0 internships and thus the
selected internship will not exist as well.

This culminates in an IndexOutOfBoundsException when checking for
the current selected internship during check while executing edit.

Changes:
Add an isEmpty() check before .get().